% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rglobi.R
\name{get_interaction_table}
\alias{get_interaction_table}
\title{Returns all known child taxa with known interaction of specified source and target taxa on any rank.}
\usage{
get_interaction_table(source.taxon.names = list(),
  target.taxon.names = list(), interaction.type = "preysOn", skip = 0,
  limit = 100, opts = list(port = 7474))
}
\arguments{
\item{interaction.type}{kind of interaction}

\item{opts}{connection parameters and other options}

\item{source.taxon.name}{list of taxon names for source}

\item{target.taxon.name}{list of taxon names for target}
}
\value{
table of matching source, target and interaction types
}
\description{
Returns all known child taxa with known interaction of specified source and target taxa on any rank.
}
\examples{
get_interaction_table(source.taxon.names = list("Aves"), target.taxon.names = list('Insecta'), interaction.type = 'preysOn')
}
\seealso{
Other interactions: \code{\link{get_child_taxa}};
  \code{\link{get_interaction_matrix}};
  \code{\link{get_interaction_types}};
  \code{\link{get_interactions_by_taxa}};
  \code{\link{get_interactions}};
  \code{\link{get_predators_of}}; \code{\link{get_prey_of}}
}

