% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglobi.R
\name{get_predators_of}
\alias{get_predators_of}
\title{Get a List of Predators of a Given Prey Taxon}
\usage{
get_predators_of(taxon = "Rattus rattus", ...)
}
\arguments{
\item{taxon}{scientific name of prey taxon. Can be any taxonomic rank (e.g. Rattus rattus, Decapoda)}

\item{...}{list of named options to configure the GloBI API}
}
\value{
list of recorded prey-predator interactions that involve the desired prey taxon.
}
\description{
Get a List of Predators of a Given Prey Taxon
}
\examples{
\donttest{
get_predators_of("Rattus rattus")
get_predators_of("Primates")
}
}
\seealso{
Other interactions:
\code{\link[=get_interaction_matrix]{get_interaction_matrix()}},
\code{\link[=get_interaction_types]{get_interaction_types()}},
\code{\link[=get_interactions]{get_interactions()}},
\code{\link[=get_interactions_by_taxa]{get_interactions_by_taxa()}},
\code{\link[=get_interactions_by_type]{get_interactions_by_type()}},
\code{\link[=get_prey_of]{get_prey_of()}}
}
\concept{interactions}
