% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetInteractions}
\alias{GetInteractions}
\title{Get Species Interaction from GloBI}
\usage{
GetInteractions(taxon = "Homo sapiens", interactionType = "preysOn")
}
\arguments{
\item{taxon}{canonical scientic name of source taxon (e.g. Homo sapiens)}

\item{interactionType}{the preferred interaction type (e.g. preysOn)}
}
\value{
species interactions between source and target taxa
}
\description{
Get Species Interaction from GloBI
}
\examples{
GetInteractions("Homo sapiens", "preysOn")
GetInteractions("Insecta", "parasiteOf")
}

