% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetPredatorsOf}
\alias{GetPredatorsOf}
\title{Get a List of Predators of a Given Prey Taxon}
\usage{
GetPredatorsOf(taxon = "Rattus rattus")
}
\arguments{
\item{taxon}{scientific name of prey taxon. Can be any taxonomic rank (e.g. Rattus rattus, Decapoda)}
}
\value{
list of recorded prey-predator interactions that involve the desired prey taxon.
}
\description{
Get a List of Predators of a Given Prey Taxon
}
\examples{
GetPredatorsOf("Rattus rattus")
GetPredatorsOf("Primates")
}

