\name{globi-package}
\alias{globi-package}
\alias{globi}
\docType{package}
\title{
R access to GloBI Database
}
\description{
This package provides a set of functions to access interaction data stored in the Global Biotic Interactions (GloBI) database, whihc work with the Encyclopaedia of Life
to normalize and integrate existing datasets of species interactions.  
}
\details{
\tabular{ll}{
Package: \tab globi\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-06-02\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
J. Stephen Gosnell, Jorritt Poelen
Maintainer: Who to complain to <yourfault@somewhere.net>
J. Stephen Gosnell (stephengosnell@gmail.com)}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\keyword{ database }
\seealso{
}
\examples{
}
