% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglobi.R
\name{query}
\alias{query}
\title{Executes a Cypher Query Against GloBI's Neo4j Instance}
\usage{
query(cypherQuery, opts = list())
}
\arguments{
\item{cypherQuery}{Cypher query (see http://github.com/jhpoelen/eol-globi-data/wiki/cypher for examples)}

\item{opts}{list of named options to configure GloBI API}
}
\value{
result of cypher query string
}
\description{
Executes a Cypher Query Against GloBI's Neo4j Instance
}
