% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rglobi.R
\name{get_interaction_matrix}
\alias{get_interaction_matrix}
\title{Get Interaction Matrix. Constructs an interaction matrix indicating whether source taxa (rows) or target taxa (columns) are known to interact with given type.}
\usage{
get_interaction_matrix(source.taxon.names = list("Homo sapiens"),
  target.taxon.names = list("Mammalia"), interaction.type = "preysOn",
  opts = list(port = 7474))
}
\arguments{
\item{interaction.type}{the preferred interaction type (e.g. preysOn)}

\item{opts}{list of options to configure GloBI API}

\item{list}{of source taxon names (e.g. list('Mammalia', 'Aves', 'Ariopsis felis'))}

\item{list}{of target taxon names}
}
\value{
matrix representing species interactions between source and target taxa
}
\description{
Get Interaction Matrix. Constructs an interaction matrix indicating whether source taxa (rows) or target taxa (columns) are known to interact with given type.
}
\examples{
get_interaction_matrix("Homo sapiens", "Mammalia", "preysOn")
}
\seealso{
Other interactions: \code{\link{get_child_taxa}};
  \code{\link{get_interaction_table}};
  \code{\link{get_interaction_types}};
  \code{\link{get_interactions_by_taxa}};
  \code{\link{get_interactions}};
  \code{\link{get_predators_of}}; \code{\link{get_prey_of}}
}

