% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_interaction_table}
\alias{get_interaction_table}
\title{Returns all known child taxa with known interaction of specified taxa and rank.}
\usage{
get_interaction_table(source.taxon.names = list("Homo sapiens"),
  target.taxon.name = list("Aves"), interaction.type = "preysOn",
  opts = list(skip = 0, limit = 20, port = 7474))
}
\arguments{
\item{target.taxon.name}{list of taxon names for target}

\item{interaction.type}{kind of interaction}

\item{opts}{connection parameters and other options}

\item{source.taxon.name}{list of taxon names for source}
}
\value{
table of matching source, target and interaction types
}
\description{
Returns all known child taxa with known interaction of specified taxa and rank.
}
\examples{
get_interaction_table(list("Aves"))
}
\seealso{
Other interactions: \code{\link{get_child_taxa}};
  \code{\link{get_interaction_matrix}};
  \code{\link{get_interaction_types}};
  \code{\link{get_interactions_by_taxa}};
  \code{\link{get_interactions}};
  \code{\link{get_predators_of}}; \code{\link{get_prey_of}}
}

