% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglobi.R
\name{get_interaction_types}
\alias{get_interaction_types}
\title{List interactions identified in GloBI database}
\usage{
get_interaction_types(opts = list(), read_csv = read_csv_online)
}
\arguments{
\item{opts}{list of named options to configure GloBI API}

\item{read_csv}{function used to find csv associated to query url, defaulting to online query method}
}
\value{
Returns data frame of supported interaction types
}
\description{
Returns data frame with supported interaction types
}
\examples{
\dontrun{
get_interaction_types()
}
}
\seealso{
Other interactions: 
  \code{\link{get_interaction_matrix}},
  \code{\link{get_interactions_by_taxa}},
  \code{\link{get_interactions_by_type}},
  \code{\link{get_interactions}},
  \code{\link{get_predators_of}}, 
  \code{\link{get_prey_of}}
}
\concept{interactions}
\keyword{database}
