% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_interaction_types}
\alias{get_interaction_types}
\title{List interactions identified in GloBI database}
\usage{
get_interaction_types()
}
\value{
Returns data frame of supported interaction types
}
\description{
Returns data frame with supported interaction types
}
\examples{
get_interaction_types()
}
\keyword{database}

