% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_interactions}
\alias{get_interactions}
\title{Get Species Interaction from GloBI}
\usage{
get_interactions(taxon = "Homo sapiens", interactionType = "preysOn")
}
\arguments{
\item{taxon}{canonical scientic name of source taxon (e.g. Homo sapiens)}

\item{interactionType}{the preferred interaction type (e.g. preysOn)}
}
\value{
species interactions between source and target taxa
}
\description{
Get Species Interaction from GloBI
}
\examples{
get_interactions("Homo sapiens", "preysOn")
get_interactions("Insecta", "parasiteOf")
}
\seealso{
Other interactions: \code{\link{get_interaction_types}};
  \code{\link{get_interactions_by_taxa}};
  \code{\link{get_predators_of}}; \code{\link{get_prey_of}}
}

