% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rglobi.R
\name{get_interactions_by_taxa}
\alias{get_interactions_by_taxa}
\title{Return interactions involving specific taxa}
\usage{
get_interactions_by_taxa(sourcetaxon, targettaxon = NULL,
  interactiontype = NULL, accordingto = NULL,
  showfield = c("source_taxon_external_id", "source_taxon_name",
  "source_taxon_path", "source_specimen_life_stage", "interaction_type",
  "target_taxon_external_id", "target_taxon_name", "target_taxon_path",
  "target_specimen_life_stage", "latitude", "longitude", "study_citation",
  "study_external_id", "study_source_citation"), otherkeys = NULL,
  bbox = NULL, returnobservations = F, opts = list(),
  read_csv = read_csv_online)
}
\arguments{
\item{sourcetaxon}{Taxa of interest (consumer, predator, parasite); may be
specified as "Genus species" or higher level (e.g., Genus, Family, Class).}

\item{targettaxon}{Taxa of interest (prey, host); may be specified as "Genus
species" or higher level (e.g., Genus, Family, Class)}

\item{interactiontype}{Interaction types of interest (prey, host); may be
specified as listed by get_interaction_types()}

\item{accordingto}{Data source of interest}

\item{showfield}{Data fields of interest (e. g. source_taxon_external_id, source_taxon_name);
may be specified as listed by get_data_fields()}

\item{otherkeys}{list of key-value pairs to query any field not covered by other parameters;
keys may be specified as listed by get_data_fields()}

\item{bbox}{Coordinates in EPSG:4326 decimal degrees defining "left, bottom,
right, top" of bounding box}

\item{returnobservations}{if true, all individual observations are returned,
else only distinct relationships}

\item{opts}{list of named options to configure GloBI API}

\item{read_csv}{function used to find csv associated to query url, defaulting to online query method}
}
\value{
Returns data frame of interactions
}
\description{
Returns interactions involving specific taxa.  Secondary (target)
taxa and spatial boundaries may also be set
}
\note{
For data sources in which type of interactions were not specified, the interaction is labeled "interacts_with"
}
\examples{
\dontrun{
get_interactions_by_taxa(sourcetaxon = "Rattus")
get_interactions_by_taxa(sourcetaxon = "Aves", targettaxon = "Rattus")
get_interactions_by_taxa(sourcetaxon = "Rattus rattus",
bbox = c(-67.87,12.79,-57.08,23.32))
}
}
\seealso{
Other interactions:
  \code{\link{get_interaction_matrix}},
  \code{\link{get_interactions_by_taxa}},
  \code{\link{get_interactions_by_type}},
  \code{\link{get_interactions}},
  \code{\link{get_predators_of}}, 
  \code{\link{get_prey_of}}
}
\concept{interactions}
\keyword{database}
