% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rglobi.R
\name{get_interactions_in_area}
\alias{get_interactions_in_area}
\title{Return all interactions in specified area}
\usage{
get_interactions_in_area(bbox)
}
\arguments{
\item{bbox}{Coordinates in EPSG:4326 decimal degrees defining "left, bottom,
right, top" of bounding box}
}
\value{
Returns data frame of interactions
}
\description{
Returns all interactions in data base in area specified in
arguments
}
\examples{
\dontrun{
get_interactions_in_area(bbox = c(-67.87, 12.79, -57.08, 23.32))
}
}
\seealso{
Other areas: \code{\link{get_interaction_areas}}
}
\keyword{database}

