% Generated by roxygen2 (4.0.1): do not edit by hand
\name{query}
\alias{query}
\title{Executes a Cypher Query Against GloBI's Neo4j Instance}
\usage{
query(cypherQuery)
}
\arguments{
\item{cypherQuery}{Cypher query (see http://github.com/jhpoelen/eol-globi-data/wiki/cypher for examples)}
}
\value{
result of cypher query string
}
\description{
Executes a Cypher Query Against GloBI's Neo4j Instance
}

